requirements.txt
accelerate==0.17.1
albumentations==1.1.0
boto3
chumpy
dataclasses
dearpygui==1.8.0
diffusers==0.15.0
einops
einops==0.4.1
imageio-ffmpeg==0.4.8
imageio==2.26.0
kornia==0.6
kornia>0.4.0
lpips==0.1.4
matplotlib==3.7.1
mediapipe
mesh-to-sdf==0.0.14
ninja==1.11.1
numpy==1.24.3
omegaconf==2.1.1
open3d
opencv_contrib_python
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pillow==9.0.1
protobuf
pudb==2019.2
pyfqmr
PyMCubes==0.1.4
pymeshfix==0.16.2
pymeshlab==2022.2.post3
PyOpenGL==3.1.5
pyrender==0.1.45
pyrr==0.10.3
pytorch-lightning==1.9.1
pyvista==0.38.5
replicate
rich==13.3.2
rtree
scikit-image
scikit-learn==1.2.2
scipy==1.9.1
setuptools==59.5.0
streamlit>=0.73.1
tensorboardX==2.6
termcolor
test-tube>=0.7.5
tetgen==0.6.2
torch-ema==0.3
torch-fidelity==0.3.0
torchmetrics==0.6.0
tqdm==4.62.3
transformers==4.27.1
trimesh==3.20.2
xatlas==0.0.7
yacs==0.1.8
git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
git+https://github.com/openai/CLIP.git@main#egg=clip
git+https://github.com/facebookresearch/pytorch3d.git@v0.7.1
# pip install kaolin==0.13.0 -f https://nvidia-kaolin.s3.us-east-2.amazonaws.com/torch-1.13.0_cu116.html